refactor(компонент кнопки): перенесла страницу about в main, поправила ссылки
